Gave hunter some basic waypoint AI
Because of my plan to implement holes in the platform, the waypoint AI
code likely won't be in the final application (unless I add enemies that
can fly). I plan to replace the hunter's AI code with code that uses the
A* algorithm to help the hunters maneuver around the holes.