Moving on from fixing the tests
The issue seems to have something to do with Visual Studio, unique
pointers, and inheritance. More specifically, the issue doesn't occur if
WallTile has no virtual methods. (I suspect the issue is with Visual Studio,
because the problem doesn't occur with g++. I managed to isolate the
problem with a 30-line separate project.) I posted the problem on
StackOverflow and am awaiting an answer (although I doubt I'll get one for
a while).