Found bug in comparator struct
This was the bug that led to problems. Thus, the tests all now pass.
However, I'm not sure why that bug led to a heap property violation, so I
will have to look into this, in case fixing that bug simply led to an
avoidance of the larger bug that led to the heap property violation.

Also, for whatever reason, I didn't already have the comparator struct
committed to Git in the first place.